ZD-513: Install event filter to ignore Alt-F4.
By installing an event filter, we can catch the user's press of the Alt
key. However, there is no way to trap the actual Alt-F4 press or to
catch the user's release of the Alt key. So we set a timer for 5 seconds
as soon as the Alt key is pressed. If a closeEvent is received within
that time, it is ignored. So the user could still exit the application
by pressing the Alt key, holding it for 5 seconds, then pressing F4.
Closing the application by pressing Ctrl-Alt-Del and ending the process
will also work.

Here is where I found example code for this situation:
http://qt-project.org/faq/answer/how_can_i_catch_altf4_in_my_qt_application